requirements.txt
azure-core==1.24.1
azure-eventhub==5.10.0
azure-iot-device==2.11.0
azure-iot-hub==2.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
deprecation==2.1.0
idna==3.3
isodate==0.6.1
janus==1.0.0
msrest==0.7.0
oauthlib==3.2.0
packaging==21.3
paho-mqtt==1.6.1
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PySocks==1.7.1
requests==2.28.0
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
six==1.16.0
typing_extensions==4.2.0
uamqp==1.5.3
urllib3==1.26.9
